requirements.txt
Flask==2.0.1
Flask-WTF==0.15.1
requests==2.26.0
beautifulsoup4==4.9.3
pandas==1.3.3
python-dotenv==0.19.0
selenium==4.1.0
webdriver-manager==3.5.2
Pillow==8.4.0 